test(145642): Corrige nomeclatura de testes
